scripts/train.py
#!/usr/bin/env python3
"""Training script for SAC agent."""

import sys
import yaml
from pathlib import Path

# Add src to path
sys.path.append(str(Path(__file__).parent.parent / "src"))

def main():
    print("SAC Trading Agent Training")
    print("="*40)
    
    # Load config
    config_path = Path("config/training_config.yaml")
    if config_path.exists():
        with open(config_path) as f:
            config = yaml.safe_load(f)
        print("Configuration loaded successfully!")
        print(f"Episodes: {config['training']['episodes']}")
        print(f"Batch size: {config['training']['batch_size']}")
    else:
        print("Configuration file not found!")
        return
    
    # TODO: Import and run your original training code here
    print("\nNOTE: Replace this section with your original training loop from paste.txt")
    print("The repository structure is ready - just paste your code into the appropriate files!")

if __name__ == "__main__":
    main()