requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0
python-dateutil>=2.8.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0